refactor(program): simplify after duplication and dead-code review
- Delete unused init_header (duplicate of Header::init)
- Use existing get_token_account_owner helper in fixed/recurring transfer paths
- Add SECS_PER_HOUR const + PlanTerms::period_length_secs helper; replace 3600 magic numbers
- Remove redundant plan_bump check in subscribe (verify_plan_pda already enforces bump)
- Single-pass Plan::check_destination (drop iterator clone)
- Move state/versioning logic from mod.rs into core.rs (mod.rs convention)
- Extract resolve_optional_payer helper in helpers/system.rs (reused 3x)
- Unify FixedTransferAccounts and RecurringTransferAccounts into shared DelegationTransferAccounts
- Strip narrating WHAT-comments in transfer/cancel handlers; keep WHY-comments

No behavior change; no IDL drift. All 38 unit tests pass under cargo test-sbf;
clippy -D warnings clean for subscriptions crate.
