+++ program/src/instructions/cancel_subscription.rs
@@ -32,37 +32,29 @@ pub fn process(accounts: &[AccountView]) -> ProgramResult {
         check_and_update_version(&mut binding)?;
         let subscription = SubscriptionDelegation::load_mut_with_min_size(&mut binding)?;
 
-        // Verify caller is the subscriber (delegator)
         if subscription.header.delegator != *accounts_struct.subscriber.address() {
             return Err(SubscriptionsError::Unauthorized.into());
         }
 
-        // Check not already cancelled
         if subscription.expires_at_ts != 0 {
             return Err(SubscriptionsError::SubscriptionAlreadyCancelled.into());
         }
 
-        // Validate subscription's delegatee matches the passed plan_pda
         if subscription.header.delegatee != *accounts_struct.plan_pda.address() {
             return Err(SubscriptionsError::SubscriptionPlanMismatch.into());
         }
 
         plan_pda = subscription.header.delegatee;
 
-        // Compute expires_at_ts based on plan state
         if accounts_struct.plan_pda.owned_by(&crate::ID) {
-            // Plan is valid — load it and verify terms match
             let plan_data = accounts_struct.plan_pda.try_borrow()?;
             let plan = Plan::load(&plan_data)?;
 
             if subscription.check_plan_terms(&plan.data.terms).is_err() {
-                // Plan terms mismatch (ghost plan) — expire immediately
+                // Ghost plan (terms mismatch) — expire immediately so the subscriber can revoke without paying.
                 expires_at_ts = current_ts;
             } else {
-                // Terms match — compute end of current period
-                let period_length_s = (subscription.terms.period_hours as i64)
-                    .checked_mul(3600)
-                    .ok_or::<ProgramError>(SubscriptionsError::ArithmeticOverflow.into())?;
+                let period_length_s = subscription.terms.period_length_secs() as i64;
                 let period_start = subscription.current_period_start_ts;
                 let elapsed = current_ts.saturating_sub(period_start);
                 let periods_elapsed = elapsed / period_length_s;
@@ -70,19 +62,18 @@ pub fn process(accounts: &[AccountView]) -> ProgramResult {
                     .checked_add(1)
                     .and_then(|p| p.checked_mul(period_length_s))
                     .and_then(|offset| period_start.checked_add(offset))
-                    // Cap at plan end so subscriber can revoke as soon as the plan expires
+                    // Cap at plan end so the subscriber can revoke as soon as the plan expires.
                     .map(|ts| if plan.data.end_ts != 0 { ts.min(plan.data.end_ts) } else { ts })
                     .ok_or::<ProgramError>(SubscriptionsError::ArithmeticOverflow.into())?;
             }
         } else {
-            // Plan is closed (not owned by our program) — expire immediately
+            // Plan account closed — expire immediately.
             expires_at_ts = current_ts;
         }
 
         subscription.expires_at_ts = expires_at_ts;
     }
 
-    // Emit SubscriptionCancelled event via self-CPI
     let event = SubscriptionCancelledEvent::new(plan_pda, *accounts_struct.subscriber.address(), expires_at_ts);
     let event_data = event.to_bytes();
     event_engine::emit_event(&crate::ID, accounts_struct.event_authority, accounts_struct.self_program, &event_data)?;
