+++ program/src/instructions/transfer_fixed_delegation.rs
@@ -1,18 +1,18 @@
 use pinocchio::{
-    error::ProgramError,
     sysvars::{clock::Clock, Sysvar},
     AccountView, Address, ProgramResult,
 };
 
 use crate::{
     check_and_update_version,
-    constants::{TOKEN_ACCOUNT_OWNER_END, TOKEN_ACCOUNT_OWNER_OFFSET},
     event_engine::{self, EventSerialize},
     events::FixedTransferEvent,
-    helpers::{transfer_with_delegate, validate_fixed_transfer, Delegation, TransferAccounts, TransferData},
+    helpers::{
+        get_token_account_owner, transfer_with_delegate, validate_fixed_transfer, Delegation,
+        DelegationTransferAccounts, TransferAccounts, TransferData,
+    },
     state::FixedDelegation,
-    AccountCheck, ProgramAccount, SignerAccount, SubscriptionAuthorityAccount, SubscriptionsError,
-    TokenAccountInterface, TokenProgramInterface, WritableAccount,
+    SubscriptionsError,
 };
 
 /// Instruction discriminator byte for `TransferFixed`.
@@ -24,7 +24,7 @@ pub const DISCRIMINATOR: &u8 = &4;
 /// `amount`, performs the SPL token transfer via the [`SubscriptionAuthority`](crate::SubscriptionAuthority)
 /// PDA, and emits a [`FixedTransferEvent`].
 pub fn process(accounts: &[AccountView], transfer: &TransferData) -> ProgramResult {
-    let accounts_struct = FixedTransferAccounts::try_from(accounts)?;
+    let accounts_struct = DelegationTransferAccounts::try_from(accounts)?;
 
     let remaining_amount: u64;
     let delegatee_address: Address;
@@ -34,7 +34,6 @@ pub fn process(accounts: &[AccountView], transfer: &TransferData) -> ProgramResu
         check_and_update_version(&mut binding)?;
         let delegation = FixedDelegation::load_mut(&mut binding)?;
 
-        // Fail fast: Check authorization first
         Delegation::check(&delegation.header, &transfer.delegator, accounts_struct.delegatee.address())?;
         if delegation.subscription_authority != *accounts_struct.subscription_authority.address() {
             return Err(SubscriptionsError::InvalidDelegatePda.into());
@@ -55,17 +54,10 @@ pub fn process(accounts: &[AccountView], transfer: &TransferData) -> ProgramResu
         init_id = delegation.header.init_id;
     }
 
-    // Extract receiver owner from token account data
-    let receiver_owner: Address;
-    {
+    let receiver_owner: Address = {
         let receiver_data = accounts_struct.receiver_ata.try_borrow()?;
-        if receiver_data.len() < TOKEN_ACCOUNT_OWNER_END {
-            return Err(SubscriptionsError::InvalidAccountData.into());
-        }
-        let mut owner_bytes = [0u8; 32];
-        owner_bytes.copy_from_slice(&receiver_data[TOKEN_ACCOUNT_OWNER_OFFSET..TOKEN_ACCOUNT_OWNER_END]);
-        receiver_owner = Address::from(owner_bytes);
-    }
+        get_token_account_owner(&receiver_data)?
+    };
 
     transfer_with_delegate(
         transfer.amount,
@@ -80,7 +72,6 @@ pub fn process(accounts: &[AccountView], transfer: &TransferData) -> ProgramResu
         },
     )?;
 
-    // Emit FixedTransferEvent via self-CPI
     let event = FixedTransferEvent::new(
         *accounts_struct.delegation_pda.address(),
         transfer.delegator,
@@ -94,48 +85,4 @@ pub fn process(accounts: &[AccountView], transfer: &TransferData) -> ProgramResu
     event_engine::emit_event(&crate::ID, accounts_struct.event_authority, accounts_struct.self_program, &event_data)?;
 
     Ok(())
-}
-
-/// Validated accounts for the [`TransferFixed`](crate::SubscriptionsInstruction::TransferFixed) instruction.
-pub struct FixedTransferAccounts<'a> {
-    pub delegation_pda: &'a AccountView,
-    pub subscription_authority: &'a AccountView,
-    pub delegator_ata: &'a AccountView,
-    pub receiver_ata: &'a AccountView,
-    pub token_program: &'a AccountView,
-    pub delegatee: &'a AccountView,
-    pub event_authority: &'a AccountView,
-    pub self_program: &'a AccountView,
-}
-
-impl<'a> TryFrom<&'a [AccountView]> for FixedTransferAccounts<'a> {
-    type Error = ProgramError;
-
-    fn try_from(accounts: &'a [AccountView]) -> Result<Self, Self::Error> {
-        let [delegation_pda, subscription_authority, delegator_ata, receiver_ata, token_program, delegatee, event_authority, self_program] =
-            accounts
-        else {
-            return Err(SubscriptionsError::NotEnoughAccountKeys.into());
-        };
-
-        ProgramAccount::check(delegation_pda)?;
-        WritableAccount::check(delegation_pda)?;
-        WritableAccount::check(delegator_ata)?;
-        WritableAccount::check(receiver_ata)?;
-        SubscriptionAuthorityAccount::check(subscription_authority)?;
-        TokenProgramInterface::check(token_program)?;
-        TokenAccountInterface::check_accounts_with_program(token_program, &[delegator_ata, receiver_ata])?;
-        SignerAccount::check(delegatee)?;
-
-        Ok(Self {
-            delegation_pda,
-            subscription_authority,
-            delegator_ata,
-            receiver_ata,
-            token_program,
-            delegatee,
-            event_authority,
-            self_program,
-        })
-    }
 }
